apps/api/pyproject.toml
[project]
name = "guardian-benefits-hub-v1-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "python-jose[cryptography]>=3.3.0",
  "cryptography>=42.0.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "paramiko>=3.4.0",
  "apscheduler>=3.10.0",
]

[tool.uvicorn]
factory = false

